requirements.txt
torch
torchvision
torchaudio
transformers
datasets
Pillow
numpy<2.0.0
scikit-image
tqdm
matplotlib
timm
pyyaml
scipy